requirements.txt
fastapi
uvicorn
python-dotenv
motor
pymongo
sentence-transformers
huggingface_hub
google-generativeai
pandas
PyPDF2
python-multipart
openpyxl
cloudinary
pdf2image
pytesseract
Pillow
python-jose[cryptography]
werkzeug
python-multipart